refactor: Split `upsertApplicationCommands` into two files, for guild and global commands respectively.
